Factored Percival emulator frame definitions into shared header
Created new header file PercivalEmulatorDefinitions.h, which contains
all the PERCIVAL-specific constants, header structures etc, to be shared
with filewriter.
